+++ utest/layers/relu_layer_test.cpp
@@ -42,9 +42,9 @@ void relu_bprop_cpu(const float* d_out, float* d_in, int len) {
   }
 }
 
-void relu_test(int dim0, int dim1) {
+void relu_test(size_t dim0, size_t dim1) {
   std::shared_ptr<GeneralBuffer<float>> buf(new GeneralBuffer<float>());
-  vector<int> dims = {dim0, dim1};
+  vector<size_t> dims = {dim0, dim1};
   std::shared_ptr<Tensor<float>> in_tensor(new Tensor<float>(dims, buf));
   std::shared_ptr<Tensor<float>> out_tensor(new Tensor<float>(dims, buf));
   buf->init(0);
