+++ utest/loss/multi_cross_entropy_loss_test.cpp
@@ -24,7 +24,7 @@ using namespace std;
 using namespace HugeCTR;
 using namespace HugeCTR::test;
 
-void multi_cross_entropy_loss(int label_dim, int batch_size) {
+void multi_cross_entropy_loss(size_t label_dim, size_t batch_size) {
   std::shared_ptr<GeneralBuffer<float>> input_b(new GeneralBuffer<float>());
   std::shared_ptr<GeneralBuffer<float>> label_b(new GeneralBuffer<float>());
   std::shared_ptr<GeneralBuffer<float>> loss_b(new GeneralBuffer<float>());
@@ -50,8 +50,8 @@ void multi_cross_entropy_loss(int label_dim, int batch_size) {
   float *d_label = label_b->get_ptr_with_offset(0);
   float *d_loss = loss_b->get_ptr_with_offset(0);
 
-  for (int i = 0; i < batch_size * label_dim; ++i) h_input[i] = rand() % 100 * 0.01f;
-  for (int i = 0; i < batch_size * label_dim; ++i) h_label[i] = rand() % 3 - 1;
+  for (size_t i = 0; i < batch_size * label_dim; ++i) h_input[i] = rand() % 100 * 0.01f;
+  for (size_t i = 0; i < batch_size * label_dim; ++i) h_label[i] = rand() % 3 - 1;
   cudaMemcpy(d_input, h_input.get(), sizeof(float) * batch_size * label_dim,
              cudaMemcpyHostToDevice);
   cudaMemcpy(d_label, h_label.get(), sizeof(float) * batch_size * label_dim,
@@ -72,7 +72,7 @@ void multi_cross_entropy_loss(int label_dim, int batch_size) {
 
   const float MIN_ = 1e-6;
   float cpu_loss = 0.f;
-  for (int i = 0; i < batch_size * label_dim; i++) {
+  for (size_t i = 0; i < batch_size * label_dim; i++) {
     float x = h_input[i];
     float y = h_label[i];
     float val = 1.f / (1.f + exp(-x));
